Fix bug with incorrect component registration
